requirements.txt
python-dotenv
justai
justdays
sqlalchemy
psycopg2-binary
boto3